requirements.txt
# Core framework
llama-index>=0.10.0
llama-index-vector-stores-chroma>=0.1.0

# Free local embeddings (no API key needed)
llama-index-embeddings-huggingface>=0.2.0
sentence-transformers>=2.7.0

# Reranker (second-pass relevance scoring)
llama-index-postprocessor-flag-embedding-reranker>=0.1.0

# Free local LLM via Ollama
llama-index-llms-ollama>=0.1.0

# Vector store (local, no API key)
chromadb>=0.5.0

# Document loaders
beautifulsoup4>=4.12.0
requests>=2.31.0

# Video transcription (free, local)
openai-whisper>=20231117

# Utilities
python-dotenv>=1.0.0
rich>=13.0.0        # pretty terminal output
psutil>=5.9.0       # system RAM monitoring (for diagnostics)
